+++ src/org/rrhs/asteroids/network/serverActions/StopShipAction.java
@@ -1,4 +1,4 @@
-package org.rrhs.asteroids.network.actions;
+package org.rrhs.asteroids.network.serverActions;
 
 import org.rrhs.asteroids.actors.NetworkActor;
 import org.rrhs.asteroids.network.Server;
@@ -6,23 +6,25 @@ import org.rrhs.asteroids.network.Server;
 import java.util.HashMap;
 import java.util.Map;
 
-public class TurnShipAction implements NetworkAction {
+public class StopShipAction implements NetworkAction {
     @Override
     public void act(Server server, int clientID, Map<String, String> parsedMessage) {
-        Map<String, String> messageToSend = new HashMap<>();
+        String action = parsedMessage.get("action");
         NetworkActor ship = server.getShip();
-        String direction = parsedMessage.get("direction");
-        if ("left".equals(direction))
+        Map<String, String> messageToSend = new HashMap<>();
+
+        if ("turn".equals(action))
         {
-            ship.setRotationSpeed(-1);
+            //make ship stop turning
+            ship.setRotationSpeed(0);
         } else
         {
-            ship.setRotationSpeed(1);
+            //make ship stop moving
+            ship.setSpeed(0);
         }
         messageToSend.put("action", "update");
         messageToSend.put("type", "ship");
         messageToSend.put("actor", ship.toString());
         server.send(messageToSend.toString());
-
     }
 }
