+++ src/org/rrhs/asteroids/network/serverActions/TurnShipAction.java
@@ -1,4 +1,4 @@
-package org.rrhs.asteroids.network.actions;
+package org.rrhs.asteroids.network.serverActions;
 
 import org.rrhs.asteroids.actors.NetworkActor;
 import org.rrhs.asteroids.network.Server;
@@ -6,25 +6,23 @@ import org.rrhs.asteroids.network.Server;
 import java.util.HashMap;
 import java.util.Map;
 
-public class StopShipAction implements NetworkAction {
+public class TurnShipAction implements NetworkAction {
     @Override
     public void act(Server server, int clientID, Map<String, String> parsedMessage) {
-        String action = parsedMessage.get("action");
-        NetworkActor ship = server.getShip();
         Map<String, String> messageToSend = new HashMap<>();
-
-        if ("turn".equals(action))
+        NetworkActor ship = server.getShip();
+        String direction = parsedMessage.get("direction");
+        if ("left".equals(direction))
         {
-            //make ship stop turning
-            ship.setRotationSpeed(0);
+            ship.setRotationSpeed(-1);
         } else
         {
-            //make ship stop moving
-            ship.setSpeed(0);
+            ship.setRotationSpeed(1);
         }
         messageToSend.put("action", "update");
         messageToSend.put("type", "ship");
         messageToSend.put("actor", ship.toString());
         server.send(messageToSend.toString());
+
     }
 }
